Consolidated processing Yelp url query response data into a single function. There will be default values if parameters are not inquired, so that eligible url request string could be made even one or some of the parameters are not given by user.

+++ randomRestaurant/YelpQuery.swift
@@ -175,7 +175,7 @@ class YelpQuery: NSObject {
     
     fileprivate func filterByRating() {
         queryResults = [[String: Any]]()
-        queryResults = businesses.filter { $0["rating"] as! Double >= (queryParameters?.rating)! }
+        queryResults = businesses.filter { $0["rating"] as! Float >= (queryParameters?.rating)! }
         print("rating: \(queryParameters?.rating)")
         print("filtered: \(queryResults)")
     }
